nls: Adding accessible compose sequences for Ș and Ț (with comma below).
Compose sequences with <dead_belowcomma> exist, but very few keyboard
layouts contain that symbol.  So a more usual character is needed to be
able to easily compose Ș, ș, Ț and ț.  The semicolon is normally only
used for composing letters with ogoneks -- but only vowels take ogoneks,
so the character is free for consonants, and thus <semicolon> is used
here to compose commas below.  It is somewhat fitting, because on most
Romanian keyboards the Ș is placed on that key, and the Ț next to it.

(Oh -- the more obvious sequences with <comma> were already taken for
composing S and T with cedillas.)

Signed-off-by: Benno Schulenberg <bensberg@justemail.net>
Signed-off-by: James Cloos <cloos@jhcloos.com>