src/util/Makefile: explicitly reset LINK to not use libtool
Simply looking at libtool redefines LINK globally to use libtool, which when
you're trying to cross-compile to Windows can cause complications.

As in src/util/ we're simply building a small binary for the build host, reset
LINK to the automake default so that the traditional compile/link steps occur
without libtool.

Also remove -all-static from LDFLAGS as that is a libtool-specific argument
intended to solve this problem.

Closes: #100
Signed-off-by: Ross Burton <ross.burton@intel.com>