Braille: Fix default lookup
cbb59d172 ('Braille: Fix typing quickly') broke the default lookup that
translates Braille keysym patterns to Braille Unicode patterns since it
rightfully clears brl_committing, but then we do not have it any more to
fill brl_committed.

This change saves the committed pattern so we can return it in the
default lookup.

Signed-off-by: Samuel Thibault <samuel.thibault@ens-lyon.org>