nls: Allow all letters with a caron to be composed also with v.
It was strange that the accented letters Ž and ž can be composed with
sequences that start with "v" ("v Z" and "v z"), but not Č and č and
Š and š (and other letters with a caron).  For these letters, compose
sequences that start with a "c" had to be used, which was frustrating
because it is hard to remember that "c" stands for "caron", AND the
graphically more obvious "v" is right next to it.

(Unfortunately, the sequence "v l" is already taken for vertical line.
Maybe the compose sequences for vertical line could be reduced to just
"V L" and "L V"?)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>