docs: unwrap the comments of structure elements in the man pages
It doesn't matter that these lines get wider than 80 columns, as it's
easy enough to make the pager scroll a bit to the right, or to have a
terminal that is a 100 or more columns wide.  It looks better to not
hard-wrap these comments.

Also, fix the layout of two table elements.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>