nls: remove redundant compose sequences with the deprecated 'underbar'
The key symbol 'underbar' is equivalent to 'underscore' -- the mirror
compose sequences with the latter symbol continue to exist.

Also, no keyboard layout in the past fifteen years has contained the
symbol 'underbar' on any key -- except a Telugu layout since a few
years, but that will be corrected soon.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>