nls: remove the compose sequences with the deviant KP_Space symbol
Using this KP_Space symbol as a stand-in for the symbol 2 doesn't make
sense.  It looks like a mistake, or as if someone had a broken keyboard
and used KP_Space as a substitute for 2.

Also, no keyboard layout in the last fifteen years has contained the
KP_Space symbol, so I don't see how anyone could type it.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>