XKBCvt: remove an if that never gets triggered
The 'if' at the beginning of _XkbHandleSpecialSym() allows only symbols
from the numeric keypad and some control symbols to pass -- XK_hyphen
is not among them, so the check for XK_hyphen in the later 'if' will
always be false.

(The corresponding conversion in _XTranslateKeySym() in KeyBind.c was
removed in March 1994.)

(Also, several keyboard layouts nowadays contain the 'hyphen' symbol,
allowing the user to enter soft hyphens into a document.  So we really
don't want to remap this symbol.)

Fixes issue #48.

Reported-by: Bhavi Dhingra

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>