nls: add compose sequences for leftward, upward and downward double arrow
These sequences follow the logic of: the arrow shaft (=) first, the
arrow head (< or > or ^ or v) second.  This is so because the sequence
<= is already taken for ≤, and there is no obvious ASCII character for
a double vertical line, so we have to make do with a horizontal one.

Since commit 6101b967b6 from eight years ago, there is a compose sequence
for the rightward double arrow, but not for the other three directions.

This fixes issue #138.

Requested-by: Mélanie Chauvel

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>