nls: delete compose sequences that mix top-row digits with numpad digits
Compose sequences for circled numbers, like ⑫  or ㉑, are nice to have,
but allowing them to be composed by typing one digit on the top row and
the other on the numerical keypad (or the other way around) is over the
top.  Remove these absurd sequences.  Keep only the sequences where both
digits are either on the top row or on the numerical keypad.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>