nls: delete some pointless and wrong compose sequences meant for Bépo
Fourteen years ago, commit 7302984642 added some four hundred compose
sequences for the benefit of the French Bépo layout.  But among these
four hundred there are several that use symbols that are not available
in the Bépo layout and are thus impossible to type.  Some of the used
symbols, like Ahook, Ehook and Ohook, are not even present in *any*
layout, making these sequences a dead weight in the Compose file.

The Amacron and Omacron are available only in the Latvian, Hawaiian,
and Maori layouts, and the Omacron also in the Silesian layout.  But
the Latvian layouts and the Hawaiian do not contain any dead keys.
Only in the Maori and Silesian layouts these sequences with Amacron
and Omacron could be typed, but that was not why they were added.

More importantly, as James Cloos noted in issue #54, sequences like
`<dead_abovedot> <amacron>` for generating `ǡ` (that is: the macron
above the dot) are questionable, as in compose sequences generally
the first accent typed is the uppermost in the composed character.

Reference:
  https://gitlab.freedesktop.org/xorg/lib/libx11/-/issues/54#note_17321

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>