Remove KOI8-R character set from en_US.UTF-8/XLC_LOCALE
https://gitlab.freedesktop.org/xorg/lib/libx11/-/blob/master/src/xlibi18n/lcCT.c#L58
has no escape sequence for KOI8-R.
That makes Xutf8TextListToTextProperty() sometimes fail when trying to
convert to COMPOUND_TEXT and when the KOI8-R charset is tried.

This is the cause for this bug:

https://github.com/ibus/ibus/issues/2422

Removing the KOI8-R charset entry fixes the problem.

This commit also fixes a few wrong texts in comments.

Signed-off-by: Mike FABIAN <mfabian@redhat.com>