nls: delete two compose sequences with an anomalous post-fixed breve
Two years ago, commit b126bfd7fe allowed using also a lowercase `u`
wherever an uppercase `U` was used to represent a breve.  But the
commit should have limited itself to only the prefixed uses of `U`,
as that is how most letters with a breve are composed.

Also, group the two compose sequences with an uppercase post-fixed `U`
together with the corresponding other post-fixed sequences.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>